+++ src/app/services/ai.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from "@angular/core";
-import { AiRespose } from "../models/ai-response.model";
 import { WtWood } from "../models/wt-wood";
 import { Failure } from "../utils/failure.utils";
 import * as tf from "@tensorflow/tfjs";
+import { ESPECIES } from "src/assets/data/especies";
 
 /**
  * Failure for Ai Domain.
@@ -23,7 +23,7 @@ export class AiService {
   /**
    * Modelo de la AI, que se carga en el constructor.
    */
-  public model: tf.LayersModel; // tf.Model
+  public model: tf.LayersModel;
 
   constructor() {
     this.loadModel();
@@ -34,7 +34,6 @@ export class AiService {
    */
   private async loadModel() {
     this.model = await tf.loadLayersModel("/assets/tf/model.json");
-    console.log(this.model.summary());
   }
 
   /**
@@ -44,93 +43,38 @@ export class AiService {
    * @param wood
    * @returns
    * @throws AiFailure.
+   *
+   * @dev Para quitar el fondo negro de la imágenes puede usarse un proceso llamado Binarización.
    */
   async withLocalImage(wood: WtWood): Promise<WtWood> {
-    // console.log("image");
-    // console.log(image);
-
-    // const canvas = document.createElement("canvas");
-    // const ctx = canvas.getContext("2d");
-    // ctx.drawImage(image, 0, 0);
-
     // Load image in a new img element.
     const image = new Image();
     image.src = wood.url;
-
-    // Build a 4d tensor with image and 3 channels.
+    // Build a 3d tensor with image 3 dimensions (width, heigth, color chanels).
     const imageTensor = tf.browser.fromPixels(image, 3);
-
-    // Resize image to 50x50.
+    // Resize image to 50x50 (size used to train model).
     const resizedTensor = tf.image.resizeBilinear(imageTensor, [50, 50]);
-
-    /**
-     * @dev Binarización de imagenes para quitar el fondo negro.
-     */
-
     // Cast tensor to float32.
     const castedTensor = tf.cast(resizedTensor, "float32");
-
-    // Divide tensor by 255 to work with numbers from 0 - 255.
+    // Divide tensor by 255 to work with smaller numbers (0 to 1).
     const dividedTensor = castedTensor.div(tf.scalar(255.0));
+    // Expand tensor in 1 dimension. Model can take a 4th dimension to analyze more than 1 image.
+    const tensor = dividedTensor.expandDims();
 
-    // Expand tensor in 1 dimension.
-    const finalTensor = dividedTensor.expandDims();
+    // Make prediction.
+    const result = (await (
+      this.model.predict(tensor) as tf.Tensor<tf.Rank>
+    ).data()) as Float32Array;
 
-    // const imageTensor4d = imageTensor.as4D(1, 50, 50, 1);
+    // Extract matchValue and matchIndex.
+    const matchValue = Math.max(...result);
+    const matchIndex = result.indexOf(matchValue);
+    const acierto = matchValue;
+    const especieResultante = ESPECIES.find(
+      (e) => e.codigo === matchIndex
+    ).nombreCientifico;
 
-    // Predict.
-    // const result = (
-    //   this.model.predict(imageTensor4d) as tf.Tensor<tf.Rank>
-    // ).dataSync();
-
-    const result = (
-      this.model.predict(finalTensor) as tf.Tensor<tf.Rank>
-    ).data();
-
-    console.log("result");
-    console.log(result);
-
-    //  0 Guazuma_ulmifolia
-    //  1 Tectona_grandis
-    //  2 Erythrina_sp
-    //  3 Peltogyne_purpurea
-    //  4 Cedrela_odorata
-    //  5 Anacardium_excelsum
-    //  6 Cordia_alliodora
-    //  7 Couma_macrocarpa
-    //  8 Eucalyptus_tereticornis
-    //  9 Pinus_radiata
-    //  10 Centrolobium_yavizanum
-    //  11 Ocotea_bofo
-    //  12 Albizia_guachapele
-    //  13 Tabebuia_rosea
-    //  14 Xylosmaz_benthamii
-    //  15 Melicoccus_bijugatus
-    //  16 Ficus_sp
-    //  17 Pithecellobium_dulce
-    //  18 Jacaranda_copaia
-    //  19 Pinus_patula
-    //  20 Gmelina_arborea
-    //  21 Qualea_acuminata
-    //  22 Clathrotropis_brunnea
-    //  23 Ocotea_insularis
-    //  24 Quercus_humboldtii
-    //  25 Campnosperma_panamensis
-    //  26 Mangifera_indica
-    //  27 Cedrelinga_cateniformis
-    //  28 Cariniana_pyriformis
-
-    // const img = tf.browser.fromPixels(wood.url);
-
-    await new Promise((resolve) => setTimeout(resolve, 5000));
-
-    /**
-     * @todo @mario Aquí se debe consumir la API de la AI, con la imagen que está en wood.localPath.
-     * Y capturar los errore y emitirlos en caso de falla.
-     *
-     * Además, poblar apropiadamente el Wood modificado.
-     */
-    return { ...wood, especie: "Eucalipto", acierto: 0.9 };
+    return { ...wood, especie: `${matchIndex}`, especieResultante, acierto };
   }
 
   /**
@@ -140,24 +84,11 @@ export class AiService {
    * @param wood
    * @returns
    * @throws AiFailure.
-   *
-   * @todo @mario Tener en cuenta que en este caso, la imagen solamente está en el dispositivo y no
-   * se ha subido al Storage. Por lo tanto no hay una URL útil que la AI pueda consumir. Si la
-   * implementación es así, será necesario resolver la lógica para que se suba la imagen, así sea
-   * temporalmente, y se analice.
    */
   async withRemoteImage(wood: WtWood): Promise<WtWood> {
     /**
-     * @todo @mario Eliminar Timeout. Esto solo sirve para simular que el análisis toma tiempo.
+     * @dev No se implementa porque el análisis de la imagen se hace local.
      */
-    await new Promise((resolve) => setTimeout(resolve, 500));
-
-    /**
-     * @todo @mario Aquí se debe consumir la API de la AI, con la imagen que está en wood.localPath.
-     * Y capturar los errore y emitirlos en caso de falla.
-     *
-     * Además, poblar apropiadamente el Wood modificado.
-     */
-    return { ...wood, especie: "Eucalipto", acierto: 0.9 };
+    throw new Error("Method not implemented.");
   }
 }